refactor(accounting): extract ReceivablesReportService from AccountingService (Wave-4 P4)
Behavior-preserving: getAgingReport + getBadDebtReport moved verbatim into a new
ReceivablesReportService; AccountingService delegates (signatures + controller
wiring unchanged). Aging bucket labels moved byte-for-byte (owner-pending
canonicalization untouched). Read-only, no shared helpers. Existing tests
(incl. Wave-4 P0 characterization) pass via the delegating facade.
